refactor install scripts (#362)

+++ src/assets/scripts/install-unfurl.sh
@@ -3,16 +3,23 @@ set -e
 
 VENV_DIR=${VENV_DIR:-~/unfurl}
 
+# Ensure that git is installed
+# (dependency for Unfurl)
+if ! which git &>/dev/null; then
+    echo "\"git\" not installed"
+    exit 1
+fi
+
 # Ensure that pip is installed
 if ! which pip &>/dev/null; then
     echo "\"pip\" not installed"
-    exit 0
+    exit 1
 fi
 
 # Ensure that python3 is installed
 if ! which python3 &>/dev/null; then
     echo "\"python3\" not installed"
-    exit 0
+    exit 1
 fi
 
 # Ensure that venv dir is set
@@ -26,7 +33,14 @@ mkdir -p "${VENV_DIR}"
 cd "${VENV_DIR}"
 python3 -m venv .venv && . .venv/bin/activate
 
+# Check that venv has been created
+# (somehow an error when creating the venv does not stop the script)
+if [ ! -f ".venv/bin/activate" ]; then
+  exit 1
+fi
+
 # Install Unfurl
 # See also https://github.com/onecommons/unfurl/blob/main/full-requirements.txt
-pip install unfurl==1.0.0 openstacksdk==0.61 python-openstackclient==6.0.0 ansible==4.10.0 pymysql==1.1.0 kubernetes==24.2.0 openshift==0.13.2 docker[tls]
+# requests==2.31.0 is required, see https://github.com/ansible-collections/community.docker/issues/868
+pip install unfurl==1.0.0 openstacksdk==0.61 python-openstackclient==6.0.0 ansible==4.10.0 pymysql==1.1.0 kubernetes==24.2.0 openshift==0.13.2 docker[tls] requests==2.31.0
 ansible-galaxy collection install community.docker:3.9.0
